ubuntu-optimizer/requirements.txt
Flask==2.0.1
requests==2.25.1
pyyaml==5.4.1
psutil==5.8.0
click==8.0.1